requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
google-generativeai>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0
roboflow>=1.1.0
chromadb
torch
torchvision
numpy
scikit-learn
ultralytics
google-generativeai
pyyaml
# Optional
flask>=3.0.0
streamlit>=1.28.0